+++ src/test/java/org/jolokia/docker/maven/util/AuthConfigHandlerTest.java
@@ -35,6 +35,8 @@ public class AuthConfigHandlerTest {
 
     private AuthConfigFactory factory;
 
+    private boolean isPush = true;
+
     public static final class MockSecDispatcher extends MockUp<SecDispatcher> {
         @Mock
         public String decrypt(String password) {
@@ -57,21 +59,21 @@ public class AuthConfigHandlerTest {
 
     @Test
     public void testEmpty() throws Exception {
-        assertNull(factory.createAuthConfig(null, settings, null, "blubberbla:1611"));
+        assertNull(factory.createAuthConfig(isPush, null, settings, null, "blubberbla:1611"));
     }
 
     @Test
     public void testSystemProperty() throws Exception {
-        System.setProperty("docker.username","roland");
-        System.setProperty("docker.password", "secret");
-        System.setProperty("docker.email", "roland@jolokia.org");
+        System.setProperty("docker.push.username","roland");
+        System.setProperty("docker.push.password", "secret");
+        System.setProperty("docker.push.email", "roland@jolokia.org");
         try {
-            AuthConfig config = factory.createAuthConfig(null,settings, null, null);
+            AuthConfig config = factory.createAuthConfig(true, null, settings, null, null);
             verifyAuthConfig(config,"roland","secret","roland@jolokia.org");
         } finally {
-            System.clearProperty("docker.username");
-            System.clearProperty("docker.password");
-            System.clearProperty("docker.email");
+            System.clearProperty("docker.push.username");
+            System.clearProperty("docker.push.password");
+            System.clearProperty("docker.push.email");
         }
     }
 
@@ -92,7 +94,7 @@ public class AuthConfigHandlerTest {
         executeWithTempHomeDir(new HomeDirExecutor() {
             @Override
             public void exec(File dir) throws IOException, MojoExecutionException {
-                AuthConfig config = factory.createAuthConfig(null, settings, "roland", null);
+                AuthConfig config = factory.createAuthConfig(isPush, null, settings, "roland", null);
                 assertNull(config);
             }
         });
@@ -114,10 +116,10 @@ public class AuthConfigHandlerTest {
         void exec(File dir) throws IOException, MojoExecutionException;
     }
 
-    private void checkDockerLogin(File homeDir,String configRegistry, String lookupRegistry) throws IOException,
-                                                                                    MojoExecutionException {
+    private void checkDockerLogin(File homeDir,String configRegistry, String lookupRegistry)
+            throws IOException, MojoExecutionException {
         createDockerConfig(homeDir, "roland", "secret", "roland@jolokia.org", configRegistry);
-        AuthConfig config = factory.createAuthConfig(null, settings, "roland", lookupRegistry);
+        AuthConfig config = factory.createAuthConfig(isPush, null, settings, "roland", lookupRegistry);
         verifyAuthConfig(config,"roland","secret","roland@jolokia.org");
     }
 
@@ -151,7 +153,7 @@ public class AuthConfigHandlerTest {
                 createOpenShiftConfig(homeDir,"openshift_simple_config.yaml");
                 Map<String,String> authConfigMap = new HashMap<>();
                 authConfigMap.put("useOpenShiftAuth","true");
-                AuthConfig config = factory.createAuthConfig(authConfigMap, settings, "roland", null);
+                AuthConfig config = factory.createAuthConfig(isPush, authConfigMap, settings, "roland", null);
                 verifyAuthConfig(config,"admin","token123",null);
             }
         });
@@ -165,7 +167,7 @@ public class AuthConfigHandlerTest {
                 @Override
                 public void exec(File homeDir) throws IOException, MojoExecutionException {
                     createOpenShiftConfig(homeDir, "openshift_simple_config.yaml");
-                    AuthConfig config = factory.createAuthConfig(null, settings, "roland", null);
+                    AuthConfig config = factory.createAuthConfig(isPush, null, settings, "roland", null);
                     verifyAuthConfig(config, "admin", "token123", null);
                 }
             });
@@ -184,7 +186,7 @@ public class AuthConfigHandlerTest {
                     createOpenShiftConfig(homeDir, "openshift_simple_config.yaml");
                     Map<String,String> authConfigMap = new HashMap<>();
                     authConfigMap.put("useOpenShiftAuth","true");
-                    AuthConfig config = factory.createAuthConfig(authConfigMap, settings, "roland", null);
+                    AuthConfig config = factory.createAuthConfig(isPush, authConfigMap, settings, "roland", null);
                     assertNull(config);
                 }
             });
@@ -201,7 +203,7 @@ public class AuthConfigHandlerTest {
                 createOpenShiftConfig(homeDir,"openshift_nologin_config.yaml");
                 Map<String,String> authConfigMap = new HashMap<>();
                 authConfigMap.put("useOpenShiftAuth","true");
-                AuthConfig config = factory.createAuthConfig(authConfigMap, settings, "roland", null);
+                AuthConfig config = factory.createAuthConfig(isPush, authConfigMap, settings, "roland", null);
             }
         });
 
@@ -214,12 +216,6 @@ public class AuthConfigHandlerTest {
         IOUtil.copy(getClass().getResourceAsStream(testConfig),new FileWriter(config));
     }
 
-
-    @Test(expected = MojoExecutionException.class)
-    public void testSystemPropertyNoUser() throws Exception {
-        checkException("docker.password");
-    }
-
     @Test(expected = MojoExecutionException.class)
     public void testSystemPropertyNoPassword() throws Exception {
         checkException("docker.username");
@@ -228,7 +224,7 @@ public class AuthConfigHandlerTest {
     private void checkException(String key) throws MojoExecutionException {
         System.setProperty(key, "secret");
         try {
-            factory.createAuthConfig(null, settings, null, null);
+            factory.createAuthConfig(isPush, null, settings, null, null);
         } finally {
             System.clearProperty(key);
         }
@@ -241,7 +237,19 @@ public class AuthConfigHandlerTest {
         pluginConfig.put("password", "secret");
         pluginConfig.put("email", "roland@jolokia.org");
 
-        AuthConfig config = factory.createAuthConfig(pluginConfig,settings,null,null);
+        AuthConfig config = factory.createAuthConfig(isPush, pluginConfig, settings, null, null);
+        verifyAuthConfig(config, "roland", "secret", "roland@jolokia.org");
+    }
+
+    @Test
+    public void testFromPluginConfigurationPull() throws MojoExecutionException {
+        Map pullConfig = new HashMap();
+        pullConfig.put("username", "roland");
+        pullConfig.put("password", "secret");
+        pullConfig.put("email", "roland@jolokia.org");
+        Map pluginConfig = new HashMap();
+        pluginConfig.put("pull",pullConfig);
+        AuthConfig config = factory.createAuthConfig(false, pluginConfig, settings, null, null);
         verifyAuthConfig(config, "roland", "secret", "roland@jolokia.org");
     }
 
@@ -250,13 +258,13 @@ public class AuthConfigHandlerTest {
     public void testFromPluginConfigurationFailed() throws MojoExecutionException {
         Map pluginConfig = new HashMap();
         pluginConfig.put("username", "admin");
-        factory.createAuthConfig(pluginConfig, settings, null, null);
+        factory.createAuthConfig(isPush, pluginConfig, settings, null, null);
     }
 
     @Test
     public void testFromSettingsSimple() throws MojoExecutionException {
         setupServers();
-        AuthConfig config = factory.createAuthConfig(null,settings, "roland", "test.org");
+        AuthConfig config = factory.createAuthConfig(isPush, null, settings, "roland", "test.org");
         assertNotNull(config);
         verifyAuthConfig(config, "roland", "secret", "roland@jolokia.org");
     }
@@ -264,7 +272,7 @@ public class AuthConfigHandlerTest {
     @Test
     public void testFromSettingsDefault() throws MojoExecutionException {
         setupServers();
-        AuthConfig config = factory.createAuthConfig(null,settings, "rhuss", "test.org");
+        AuthConfig config = factory.createAuthConfig(isPush, null, settings, "rhuss", "test.org");
         assertNotNull(config);
         verifyAuthConfig(config, "rhuss", "secret2", "rhuss@redhat.com");
     }
@@ -272,7 +280,7 @@ public class AuthConfigHandlerTest {
     @Test
     public void testFromSettingsDefault2() throws MojoExecutionException {
         setupServers();
-        AuthConfig config = factory.createAuthConfig(null,settings,"tanja",null);
+        AuthConfig config = factory.createAuthConfig(isPush, null, settings, "tanja", null);
         assertNotNull(config);
         verifyAuthConfig(config,"tanja","doublesecret","tanja@jolokia.org");
     }
@@ -280,7 +288,7 @@ public class AuthConfigHandlerTest {
     @Test
     public void testWrongUserName() throws MojoExecutionException {
         setupServers();
-        assertNull(factory.createAuthConfig(null, settings, "roland", "another.repo.org"));
+        assertNull(factory.createAuthConfig(isPush, null, settings, "roland", "another.repo.org"));
     }